requirements.txt
setuptools==70.0.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
deepface==0.0.93
tf-keras==2.16.0
opencv-python-headless==4.9.0.80
scipy==1.13.0
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.2
httpx==0.27.0
redis==5.0.4
celery==5.4.0
python-dotenv==1.0.1